refactor(RHOAIENG-79894): relocate shared metrics cluster into ui-core
Move the self-contained, cross-feature metrics primitives out of
frontend/src/concepts/metrics/ into packages/ui-core so both Distributed
Workloads and model-serving consume them from a shared core library
rather than through @odh-dashboard/internal.

Moved to ui-core:
- MetricsCommonContext (contexts/)
- MetricsPageToolbar, MetricsTimeRangeSelect, MetricsRefreshIntervalSelect (components/)
- useCurrentTimeframeBrowserStorage, useRefreshIntervalBrowserStorage (hooks/)
- isTimeframeTitle / isRefreshIntervalTitle guards (utilities/metrics.ts) + tests

Note: the ticket table lists model-serving as the target for the shared
context/toolbar, but these are used by Distributed Workloads too; homing
them in model-serving would create a feature->feature dependency
(package-topology.md Rule 1), so they live in ui-core instead.

Selects and browser-storage hooks stay internal to ui-core (reached via
relative imports); only MetricsPageToolbar and MetricsCommonContext are
exported. 13 consumers re-pointed to @odh-dashboard/ui-core.
