+++ frontend/src/concepts/distributedWorkloads/DistributedWorkloadsContext.tsx
@@ -5,6 +5,7 @@ import { getDisplayNameFromK8sResource } from '@odh-dashboard/k8s-core';
 import { FetchStateObject } from '@odh-dashboard/ui-core/hooks/useFetch';
 import { DEFAULT_LIST_FETCH_STATE } from '@odh-dashboard/ui-core/utilities/fetchState';
 import { RefreshIntervalValue } from '@odh-dashboard/ui-core/utilities/metrics';
+import { MetricsCommonContext } from '@odh-dashboard/ui-core/contexts/MetricsCommonContext';
 import { ClusterQueueKind, LocalQueueKind, WorkloadKind } from '#~/k8sTypes';
 import { conditionalArea } from '#~/concepts/areas';
 import useSyncPreferredProject from '#~/concepts/projects/useSyncPreferredProject';
@@ -16,7 +17,6 @@ import {
   getGenericErrorCode,
   useDWProjectCurrentMetrics,
 } from '#~/api';
-import { MetricsCommonContext } from '#~/concepts/metrics/MetricsCommonContext';
 import PermissionsNotSet from './PermissionsNotSet';
 import useClusterQueues from './useClusterQueues';
 import useLocalQueues from './useLocalQueues';
