--- frontend/src/concepts/metrics/utils.ts
@@ -1,9 +1,0 @@
-import { isEnumMember } from '@odh-dashboard/foundation';
-import { TimeframeTitle, RefreshIntervalTitle } from '@odh-dashboard/ui-core/types/metrics';
-
-export const isTimeframeTitle = (timeframe?: string | null): timeframe is TimeframeTitle =>
-  isEnumMember(timeframe, TimeframeTitle);
-
-export const isRefreshIntervalTitle = (
-  refreshInterval?: string | null,
-): refreshInterval is RefreshIntervalTitle => isEnumMember(refreshInterval, RefreshIntervalTitle);
