+++ packages/model-serving/src/components/metrics/bias/BiasTab.tsx
@@ -12,7 +12,7 @@ import {
   ToolbarItem,
 } from '@patternfly/react-core';
 import { ExclamationCircleIcon } from '@patternfly/react-icons';
-import MetricsPageToolbar from '@odh-dashboard/internal/concepts/metrics/MetricsPageToolbar';
+import MetricsPageToolbar from '@odh-dashboard/ui-core/components/MetricsPageToolbar';
 import { useModelBiasData } from '@odh-dashboard/trustyai/hooks/useModelBiasData';
 import DashboardExpandableSection from '@odh-dashboard/ui-core/components/DashboardExpandableSection';
 import { TrustyInstallState } from '@odh-dashboard/trustyai/types';
