+++ packages/model-serving/src/components/metrics/nim/NimTab.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { PageSection, Stack, StackItem } from '@patternfly/react-core';
-import MetricsPageToolbar from '@odh-dashboard/internal/concepts/metrics/MetricsPageToolbar';
+import MetricsPageToolbar from '@odh-dashboard/ui-core/components/MetricsPageToolbar';
 import type { InferenceServiceKind } from '@odh-dashboard/model-serving/shared';
 import NimMetrics from './NimMetrics';
