+++ packages/model-serving/src/components/metrics/performance/PerformanceTab.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { EmptyState, PageSection, Stack, StackItem } from '@patternfly/react-core';
 import { WarningTriangleIcon } from '@patternfly/react-icons';
 import { SupportedArea, useIsAreaAvailable } from '@odh-dashboard/plugin-core/areas';
-import MetricsPageToolbar from '@odh-dashboard/internal/concepts/metrics/MetricsPageToolbar';
+import MetricsPageToolbar from '@odh-dashboard/ui-core/components/MetricsPageToolbar';
 import type { InferenceServiceKind } from '@odh-dashboard/model-serving/shared';
 import ModelGraphs from './ModelGraphs';
