+++ packages/ui-core/package.json
@@ -40,6 +40,7 @@
     "./components/searchSelector/ProjectScopedSearchDropdown": "./src/components/searchSelector/ProjectScopedSearchDropdown.tsx",
     "./components/TruncatedText": "./src/components/TruncatedText.tsx",
     "./components/SimpleSelect": "./src/components/SimpleSelect.tsx",
+    "./components/MetricsPageToolbar": "./src/components/MetricsPageToolbar.tsx",
     "./components/TypeaheadSelect": "./src/components/TypeaheadSelect.tsx",
     "./components/NumberInputWrapper": "./src/components/NumberInputWrapper.tsx",
     "./components/FilterToolbar": "./src/components/FilterToolbar.tsx",
@@ -63,6 +64,7 @@
     "./context/ProjectHardwareProfilesContext": "./src/context/ProjectHardwareProfilesContext.tsx",
     "./context/ProjectsContext": "./src/context/ProjectsContext.tsx",
     "./context/getStoredPreferredProject": "./src/context/getStoredPreferredProject.ts",
+    "./contexts/MetricsCommonContext": "./src/contexts/MetricsCommonContext.tsx",
     "./contexts/ThemeContext": "./src/contexts/ThemeContext.tsx",
     "./contexts/AnalyticsContext": "./src/contexts/AnalyticsContext.tsx",
     "./contexts/NotificationContext": "./src/contexts/NotificationContext.tsx",
