+++ packages/ui-core/src/hooks/useRefreshIntervalBrowserStorage.ts
@@ -1,6 +1,6 @@
-import { useBrowserStorage } from '@odh-dashboard/ui-core/utilities';
-import type { SetBrowserStorageHook } from '@odh-dashboard/ui-core/utilities';
-import { RefreshIntervalTitle } from '@odh-dashboard/ui-core/types/metrics';
+import { useBrowserStorage } from './useBrowserStorage';
+import type { SetBrowserStorageHook } from './useBrowserStorage';
+import { RefreshIntervalTitle } from '../types/metrics';
 
 const useRefreshIntervalBrowserStorage = (
   initialRefreshInterval = RefreshIntervalTitle.FIVE_MINUTES,
