+++ packages/ui-core/src/utilities/__tests__/metrics.spec.ts
@@ -1,4 +1,4 @@
-import { isRefreshIntervalTitle, isTimeframeTitle } from '#~/concepts/metrics/utils';
+import { isRefreshIntervalTitle, isTimeframeTitle } from '../metrics';
 
 describe('isTimeframeTitle', () => {
   it('works in the true case', () => {
