+++ packages/ui-core/src/utilities/metrics.ts
@@ -1,3 +1,4 @@
+import { isEnumMember } from '@odh-dashboard/foundation';
 import {
   TimeframeTitle,
   RefreshIntervalTitle,
@@ -54,3 +55,10 @@ export const RefreshIntervalValue: RefreshIntervalValueType = {
  * queries (e.g. http requests) that the time series will contain values for number of requests per 5m.
  */
 export const PROMETHEUS_REQUEST_RESOLUTION = '300s';
+
+export const isTimeframeTitle = (timeframe?: string | null): timeframe is TimeframeTitle =>
+  isEnumMember(timeframe, TimeframeTitle);
+
+export const isRefreshIntervalTitle = (
+  refreshInterval?: string | null,
+): refreshInterval is RefreshIntervalTitle => isEnumMember(refreshInterval, RefreshIntervalTitle);
